+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 
 
 - [Feature: Support parse ratelimit rule expression labels.](https://github.com/Tencent/spring-cloud-tencent/pull/183)
-- [fix:Router support request label.](https://github.com/Tencent/spring-cloud-tencent/pull/165)
+- [Feature: Router support request label.](https://github.com/Tencent/spring-cloud-tencent/pull/165)
+- [Feature: Support router expression label](https://github.com/Tencent/spring-cloud-tencent/pull/190)